backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
python-multipart==0.0.12
replicate>=0.34.0
python-dotenv>=1.0.0
mercadopago>=2.4.0
pydantic[email]
filetype
